+++ src/trocr_processor.rs
@@ -1,8 +1,5 @@
-// Code shamelessely copied from
-// https://github.com/huggingface/candle/blob/main/candle-examples/examples/trocr/image_processor.rs
-// Hopefully the license lets me do this
-
-use image::{ImageBuffer, Rgb};
+use image::{ImageBuffer, Rgb, RgbImage};
+use eframe::egui::Pos2;
 
 use candle_core::{Device, Result, Tensor};
 
@@ -19,7 +16,11 @@ impl ImageProcessor {
         }
     }
 
-    pub fn preprocess(&self, image: ImageBuffer<Rgb<u8>, Vec<u8>>) -> Result<Tensor> {
+    pub fn preprocess(&self, painting_frame: &Vec<[Pos2; 2]>) -> Result<Tensor> {
+        print!("Converting painting frame to image... ");
+        let image = self.painting_frame_to_image(painting_frame);
+        println!("Image converted!");
+        
         print!("Asserting dimensions... ");
         assert_eq!(image.dimensions(), (self.width as u32, self.height as u32));
         println!("Dimensions correct!");
@@ -59,5 +60,55 @@ impl ImageProcessor {
 
         normalized_tensor
         //Err(candle_core::Error::msg("Not implemented yet."))
+    }
+    
+    // TODO: Rewrite the algorithm to not get out of bounds
+    pub fn painting_frame_to_image(
+        &self,
+        painting_frame: &Vec<[Pos2; 2]>,
+    ) -> image::ImageBuffer<Rgb<u8>, Vec<u8>> {
+        let mut img = RgbImage::new(self.width as u32, self.height as u32);
+    
+        for pixel in img.pixels_mut() {
+            *pixel = Rgb([255, 255, 255]);
+        }
+    
+        for [start, end] in painting_frame {
+            let (x0, y0) = (start.x as i32, start.y as i32);
+            let (x1, y1) = (end.x as i32, end.y as i32);
+    
+            // Bresenham's line algorithm to draw the line
+            let dx = (x1 - x0).abs();
+            let sx = if x0 < x1 { 1 } else { -1 };
+            let dy = -(y1 - y0).abs();
+            let sy = if y0 < y1 { 1 } else { -1 };
+            let mut err = dx + dy;
+    
+            let mut x = x0;
+            let mut y = y0;
+    
+            loop {
+                if x >= 0 && x < self.width as i32 && y >= 0 && y < self.height as i32 {
+                    img.put_pixel(x as u32, y as u32, Rgb([0, 0, 0]));
+                    img.put_pixel((x + 1) as u32, y as u32, Rgb([0, 0, 0]));
+                    img.put_pixel(x as u32, (y + 1) as u32, Rgb([0, 0, 0]));
+                    img.put_pixel((x + 1) as u32, (y + 1) as u32, Rgb([0, 0, 0]));
+                }
+                if x == x1 && y == y1 {
+                    break;
+                }
+                let e2 = 2 * err;
+                if e2 >= dy {
+                    err += dy;
+                    x += sx;
+                }
+                if e2 <= dx {
+                    err += dx;
+                    y += sy;
+                }
+            }
+        }
+    
+        img
     }
 }
